Accept completion with tab
Bruv TJ i tried for 2 days but it just slowed me down to do ctrl+y
And I do not see me ever wanting to insert a tab when a autocompletion
drop down is active.
I do tabs in comments or in the beginning of the line.

But I will try to remove my muscle memory to also sometimes use return
to accept completions.

Vscode accepts both tab and return.
But having the ability to insert a newline even if the editor thinks we
could also complete something is nice.