building non main go libs is more useful than running stuff in vim
Pressing "c" to compile and run is fun for small projects.
But it fails for golang projects that don't have a main package.

And running inside vim is a bit annoying for bigger projects anyways.

`go build` does the crucial part for both standalone exectuables and
libraries: it checks if the build passes